Change directory structure and fix Makefiles
The directory structure is changed to conform more with standard C
coding practices. Certain standard files (README, etc.) have been added,
but are empty for now.

The Makefile structure has been changed so that the ThinkGear files and
the project files are separate and compiled separately. The final
executable is now produced in the src directory.

A TODO is added with many necessary changes.